Code smells de ferry y booking
