+++ java/src/main/java/booking/AvailableCrossing.java
@@ -1,0 +1,12 @@
+package booking;
+
+public class AvailableCrossing {
+
+    public String ferryName;
+    public String originPort;
+    public String destinationPort;
+    public long setOff;
+    public long arrive;
+    public int seatsLeft;
+    public int journeyId;
+}
