+++ java/src/main/java/booking/Booking.java
@@ -1,0 +1,7 @@
+package booking;
+
+public class Booking {
+
+    public int journeyId;
+    public int passengers;
+}
