--- java/src/main/java/ferry/booking/AvailableCrossing.java
@@ -1,12 +1,0 @@
-package ferry.booking;
-
-public class AvailableCrossing {
-
-    public String ferryName;
-    public String originPort;
-    public String destinationPort;
-    public long setOff;
-    public long arrive;
-    public int seatsLeft;
-    public int journeyId;
-}
