--- java/src/main/java/ferry/booking/Booking.java
@@ -1,7 +1,0 @@
-package ferry.booking;
-
-public class Booking {
-
-    public int journeyId;
-    public int passengers;
-}
