--- java/src/main/java/ferry/booking/Ferry.java
@@ -1,15 +1,0 @@
-package ferry.booking;
-
-public class Ferry {
-
-    public int id;
-    public String name;
-    public int passengers;
-    public int vehicles;
-    public double weight;
-    public int homePortId;
-
-    public Ferry() {
-        homePortId = 1;
-    }
-}
