--- java/src/main/java/ferry/booking/FerryJourney.java
@@ -1,8 +1,0 @@
-package ferry.booking;
-
-public class FerryJourney {
-
-    public Ferry ferry;
-    public PortModel origin;
-    public PortModel destination;
-}
