--- java/src/main/java/ferry/booking/Port.java
@@ -1,7 +1,0 @@
-package ferry.booking;
-
-public class Port {
-
-    public int id;
-    public String name;
-}
