--- java/src/main/java/ferry/booking/TimeTable.java
@@ -1,9 +1,0 @@
-package ferry.booking;
-
-import java.util.List;
-
-public class TimeTable {
-
-    public int id;
-    public List<TimeTableEntry> entries;
-}
