--- java/src/main/java/ferry/booking/TimeTableEntry.java
@@ -1,11 +1,0 @@
-package ferry.booking;
-
-public class TimeTableEntry {
-
-    public int id;
-    public int timeTableId;
-    public int originId;
-    public int destinationId;
-    public long time;
-    public long journeyTime;
-}
