--- java/src/main/java/ferry/booking/TimeTableViewModelRow.java
@@ -1,12 +1,0 @@
-package ferry.booking;
-
-public class TimeTableViewModelRow {
-
-    public String ferryName;
-    public String destinationPort;
-    public String originPort;
-    public String startTime;
-    public String journeyLength;
-    public String arrivalTime;
-    public int journeyId;
-}
