+++ java/src/main/java/journey/FerryJourney.java
@@ -1,0 +1,11 @@
+package journey;
+
+import ferry.Ferry;
+import port.PortModel;
+
+public class FerryJourney {
+
+    public Ferry ferry;
+    public PortModel origin;
+    public PortModel destination;
+}
