+++ java/src/main/java/port/Port.java
@@ -1,0 +1,7 @@
+package port;
+
+public class Port {
+
+    public int id;
+    public String name;
+}
