+++ java/src/main/java/timetable/TimeTable.java
@@ -1,0 +1,9 @@
+package timetable;
+
+import java.util.List;
+
+public class TimeTable {
+
+    public int id;
+    public List<TimeTableEntry> entries;
+}
