+++ java/src/main/java/timetable/TimeTableEntry.java
@@ -1,0 +1,11 @@
+package timetable;
+
+public class TimeTableEntry {
+
+    public int id;
+    public int timeTableId;
+    public int originId;
+    public int destinationId;
+    public long time;
+    public long journeyTime;
+}
