+++ java/src/main/java/timetable/TimeTableViewModelRow.java
@@ -1,0 +1,12 @@
+package timetable;
+
+public class TimeTableViewModelRow {
+
+    public String ferryName;
+    public String destinationPort;
+    public String originPort;
+    public String startTime;
+    public String journeyLength;
+    public String arrivalTime;
+    public int journeyId;
+}
